feat: Implement traffic light detector
tl_detector.py uses current_pose and base_waypoints togheter with the position of the traffic lights to publish the waypoint of the next upcoming traffic light. The traffic light waypoint is being published to the traffic_waypoint topic if the upcoming traffic light is red. If the light is not red then "-1" is being published.