feat: Implement basic waypoint updater
Implemented a basic version of the `waypoint_updater` node that simply publishes a subset of the waypoints published via the `/basic_waypoints` topic without altering them in any way. This means that these published waypoints allow the car to follow the track, but it will ignore traffic lights and obstacles.