feat: Tune the throttle/braking PID controller and add a filter
Improved the PID controller tuning. Ideal tuning for the simulator is currently impossible due to an apparent simulator bug that leads to the simulator not utilizing each throttle/braking command being sent to it.

Also added a filter for smoother acceleration/deceleration. The filter weights 0.1 of the current actuation value agains 0.9 of the last actuation value.